fix: Promote BindValue to the root namespace, sharpen its docs (#282)
Sql.Bind's return type moved to BindValue last round, but the type
stayed under Internal/. That put it in tension with the root-namespace
test in public-api-design.md: the whole point of the return type is
that a caller writes "BindValue p10 = Bind(10);" to share a marker
across clauses, which is exactly the "must write its name in a
declaration position" criterion. DbColumn is the existing precedent for
a concrete, user-declared SqlExpression subtype living in the root
namespace instead of Internal/ — move BindValue to sit beside it
(src/SqlArtisan/SqlPart/Expression/), out from under the
Internal/SqlPart/** CS1591 suppression, and give it real XML docs
matching OutputParameter's house style now that it's first-class public
surface. Drop the using directives across the tree that only existed
for BindValue's old namespace; C#'s enclosing-namespace lookup already
resolves it everywhere else with no using needed. CLAUDE.md's layout
note now names both types sharing that folder.

Also fold docs/functions.md's trailing Bind cross-reference paragraph
into a parenthetical on the Bind Parameter Types intro sentence, so the
pointer to Sql.Bind lives inside a section already linked from
docs/README.md's index instead of only being reachable via the GROUP BY
clause's prose.
