+++ src/SqlArtisan/Sql/Sql.B.cs
@@ -1,4 +1,3 @@
-using SqlArtisan.Internal;
 using static SqlArtisan.Internal.ExpressionResolver;
 
 namespace SqlArtisan;
