+++ src/SqlArtisan/SqlBuilder/SqlParameters.cs
@@ -1,5 +1,3 @@
-using SqlArtisan.Internal;
-
 namespace SqlArtisan;
 
 /// <summary>
