+++ tests/SqlArtisan.Tests/FunctionTests/FunctionTests.B.cs
@@ -1,5 +1,4 @@
 using System.Text;
-using SqlArtisan.Internal;
 using static SqlArtisan.Sql;
 
 namespace SqlArtisan.Tests;
