+++ tests/SqlArtisan.Tests/ReturningTests.cs
@@ -1,6 +1,5 @@
 using System.Data;
 using System.Text;
-using SqlArtisan.Internal;
 using static SqlArtisan.Sql;
 
 namespace SqlArtisan.Tests;
